fix: /media totalSize — SUM() bigint returned as string by pg driver
SUM(integer) in Postgres returns bigint, which node-postgres gives back
as a string. UI checked typeof === 'number' → totalSize stayed 0.
- API: cast COALESCE(SUM(size),0)::int so it returns a JS number
- UI: Number(data.totalSize) with NaN guard

+++ apps/web/src/app/api/media/route.ts
@@ -82,7 +82,8 @@ export const GET = withAuth(async (ctx, req: NextRequest) => {
         }),
         db.$count(schema.media, and(...where)),
         // Total penyimpanan seluruh media org (untuk limitasi paket langganan)
-        db.select({ total: sql<number>`COALESCE(SUM(${schema.media.size}), 0)` })
+        // SUM() Postgres → bigint → pg mengembalikan string; cast ::int agar angka.
+        db.select({ total: sql<number>`COALESCE(SUM(${schema.media.size}), 0)::int` })
             .from(schema.media)
             .where(eq(schema.media.organizationId, activeOrganizationId)),
     ]);
